moved blockchain-related functions to blockchain.go
